eval/requirements.txt
requests
openai==1.7.0
python-dotenv